+++ allink_core/apps/events/abstract_models.py
@@ -8,7 +8,7 @@ from filer.fields.image import FilerImageField
 from djangocms_text_ckeditor.fields import HTMLField
 from cms.models.fields import PlaceholderField
 
-from aldryn_common.admin_fields.sortedm2m import SortedM2MModelField
+from allink_core.core.models.fields import SortedM2MModelField
 from allink_core.core.models import (
     AllinkCategoryFieldsModel,
     AllinkBaseTranslatableModel,